chore(astro): moved createMarked out of umami.ts into mdx.ts, since it's fundamentally the site's markdown-rendering function
