+++ src/utils/index.ts
@@ -12,6 +12,7 @@ export {
   ensureLeadingSlash,
   ensureAbsoluteUrl,
   getSafeExternalUrl,
+  isSafeMarkdownHref,
   getSocialIconName,
   FALLBACK_SOCIAL_ICON,
   type SocialIconName
